renderItem callback could receive the date
I don't really see why renderEmptyDate would receive it and not renderItem. I need this because my last day item is always a CTA button to "add availability". Without the date I have to resort on trying to get back the date from the items list of my component which is a bit annoying