Change months when getMonth() is different
This should fix #273 

So, actually this more the hard approach to solving that problem as it doesn't consider if hideExtraDays is disabled. In that case it should take into account of the new current date is among those disabled days. I would like to help and continue the work but I'm not sure how to know that the new date is disabled because that's different in every month.

https://github.com/wix/react-native-calendars/issues/273